Dataset: Support rolling sprint features
This option causes the features of a previous sprint to be moved to the
sprint after it, losing the first sprint of the project as a sample.
This option allows for features that contain details from the entire
sprint (not just knowledge from the start of the sprint) by passing them
to the next sprint in this way.