Initial version of analogy-based estimation
The model and accompanying runner which provides the entire training set
and labels, minus the example being trained on for automated jackknifing
while allowing the model to perform its own classification, in this case
using nearest neighbor (Euclidean) similarity to pick a top-k of analogy
examples, whose labels are then averaged for the final output. Currently
this model is not actually generative, in that it does not modify any of
its variables in order to select features, but this simple model already
seems to work fairly well, and fairly fast, using TensorFlow operations.