ABE: Output additional results from validation
Add a generic means of augmenting the validation result dictionary with
more values from the context. Models can now log additional operations
during evaluation, and the augmented results can also be transformed via
metadata that points to data set contexts. For this, the dataset now
keeps track of the original indexes for each type of set, so train and
test as well rather than just validation. Finally, the export of the
result replaces any NaN values in NumPy arrays with null for validity.